Move annotation-metadata to util directory
