+++ src/sidebar/services/root-thread.js
@@ -3,7 +3,7 @@
 const buildThread = require('../build-thread');
 const events = require('../events');
 const memoize = require('../util/memoize');
-const metadata = require('../annotation-metadata');
+const metadata = require('../util/annotation-metadata');
 const tabs = require('../tabs');
 const uiConstants = require('../ui-constants');
