+++ src/sidebar/util/annotation-metadata.js
